+++ packages/components/src/internal/util/utils.spec.ts
@@ -30,7 +30,6 @@ import {
     getDisambiguatedSelectInputOptions,
     getIconFontCls,
     getUpdatedData,
-    getUpdatedDataFromGrid,
     handleRequestFailure,
     intersect,
     isImage,
@@ -745,382 +744,6 @@ describe('getUpdatedData', () => {
             Other: 'other3',
             Data: 'data1',
         });
-    });
-});
-
-describe('getUpdatedDataFromGrid', () => {
-    const cols = Map<string, QueryColumn>({
-        rowid: new QueryColumn({ name: 'RowId', rangeURI: INTEGER_TYPE.rangeURI }),
-        value: new QueryColumn({ name: 'Value', rangeURI: INTEGER_TYPE.rangeURI }),
-        data: new QueryColumn({ name: 'Data', rangeURI: TEXT_TYPE.rangeURI }),
-        andagain: new QueryColumn({ name: 'AndAgain', rangeURI: TEXT_TYPE.rangeURI }),
-        name: new QueryColumn({ name: 'Name', rangeURI: TEXT_TYPE.rangeURI }),
-        other: new QueryColumn({ name: 'Other', rangeURI: TEXT_TYPE.rangeURI }),
-        bool: new QueryColumn({ name: 'Bool', rangeURI: BOOLEAN_TYPE.rangeURI }),
-        int: new QueryColumn({ name: 'Int', rangeURI: INTEGER_TYPE.rangeURI }),
-        date: new QueryColumn({ name: 'Date', rangeURI: DATE_TYPE.rangeURI }),
-    });
-    const queryInfo = QueryInfo.create({
-        columns: cols,
-    });
-    const queryInfoWithAltPK = QueryInfo.create({
-        columns: cols,
-        altUpdateKeys: new Set<strin>(['Data']),
-    });
-    const originalData = fromJS({
-        448: {
-            RowId: 448,
-            Value: null,
-            Data: 'data1',
-            AndAgain: 'again',
-            Name: 'S-20190516-9042',
-            Other: 'other1',
-            Bool: true,
-            Int: 0,
-            Date: '2020-12-23 14:34',
-        },
-        447: {
-            RowId: 447,
-            Value: null,
-            Data: 'data1',
-            AndAgain: 'again',
-            Name: 'S-20190516-4622',
-            Other: 'other2',
-            Bool: false,
-            Int: 7,
-            Date: null,
-        },
-        446: {
-            RowId: 446,
-            Value: 'val',
-            Data: 'data1',
-            AndAgain: 'again',
-            Name: 'S-20190516-2368',
-            Other: 'other3',
-            Bool: true,
-            Int: 6,
-            Date: '1922-08-21 00:00',
-        },
-        445: {
-            RowId: 445,
-            Value: 'val',
-            Data: 'data1',
-            AndAgain: 'again',
-            Name: 'S-20190516-9512',
-            Other: null,
-            Bool: false,
-            Int: 5,
-            Date: null,
-        },
-    });
-
-    test('no edited rows', () => {
-        const updatedData = getUpdatedDataFromGrid(originalData, [], 'RowId', queryInfo);
-        expect(updatedData).toHaveLength(0);
-    });
-
-    test('edited row with array', () => {
-        const orig = fromJS({
-            448: {
-                RowId: 448,
-                Alias: undefined,
-            },
-        });
-
-        let updatedData = getUpdatedDataFromGrid(
-            orig,
-            [
-                Map<string, any>({
-                    RowId: '448',
-                    Alias: [],
-                }),
-            ],
-            'RowId',
-            queryInfo
-        );
-        expect(updatedData).toHaveLength(0);
-
-        updatedData = getUpdatedDataFromGrid(
-            orig,
-            [
-                Map<string, any>({
-                    RowId: '448',
-                    Alias: ['test1'],
-                }),
-            ],
-            'RowId',
-            queryInfo
-        );
-        expect(updatedData).toHaveLength(1);
-    });
-
-    test('edited rows did not change', () => {
-        const updatedData = getUpdatedDataFromGrid(
-            originalData,
-            [
-                Map<string, any>({
-                    RowId: '448',
-                    Value: null,
-                    Data: 'data1',
-                    AndAgain: 'again',
-                    Name: 'S-20190516-9042',
-                    Other: 'other1',
-                    Bool: true,
-                    Int: 0,
-                    Date: '2020-12-23 14:34',
-                }),
-                Map<string, any>({
-                    RowId: '447',
-                    Value: '',
-                    Data: 'data1',
-                    AndAgain: 'again',
-                    Name: 'S-20190516-4622',
-                    Other: 'other2',
-                    Bool: false,
-                    Int: '7',
-                    Date: null,
-                }),
-                Map<string, any>({
-                    RowId: '446',
-                    Value: 'val',
-                    Data: 'data1',
-                    AndAgain: 'again',
-                    Name: 'S-20190516-2368',
-                    Other: 'other3',
-                    Bool: true,
-                    Int: '6',
-                    Date: '1922-08-21 00:00',
-                }),
-                Map<string, any>({
-                    RowId: '445',
-                    Value: 'val',
-                    Data: 'data1',
-                    AndAgain: 'again',
-                    Name: 'S-20190516-9512',
-                    Other: null,
-                    Bool: false,
-                    Int: 5,
-                    Date: null,
-                }),
-            ],
-            'RowId',
-            queryInfo
-        );
-        expect(updatedData).toHaveLength(0);
-    });
-
-    test('edited row removed values', () => {
-        const updatedData = getUpdatedDataFromGrid(
-            originalData,
-            [
-                Map<string, any>({
-                    RowId: '448',
-                    Value: null,
-                    Data: undefined,
-                    AndAgain: 'again',
-                    Name: 'S-20190516-9042',
-                    Other: 'other1',
-                    Bool: undefined,
-                    Int: undefined,
-                    Date: null,
-                }),
-                Map<string, any>({
-                    RowId: '447',
-                    Value: null,
-                    Data: 'data1',
-                    AndAgain: null,
-                    Name: 'S-20190516-4622',
-                    Other: 'other2',
-                    Bool: undefined,
-                    Int: undefined,
-                    Date: null,
-                }),
-                Map<string, any>({
-                    RowId: '446',
-                    Value: 'val',
-                    Data: 'data1',
-                    AndAgain: 'again',
-                    Name: 'S-20190516-2368',
-                    Other: 'other3',
-                    Bool: true,
-                    Int: 6,
-                    Date: '1922-08-21 00:00',
-                }),
-                Map<string, any>({
-                    RowId: '445',
-                    Value: 'val',
-                    Data: 'data1',
-                    AndAgain: 'again',
-                    Name: 'S-20190516-9512',
-                    Other: null,
-                    Bool: false,
-                    Int: 5,
-                    Date: null,
-                }),
-            ],
-            'RowId',
-            queryInfo
-        );
-        expect(updatedData).toHaveLength(2);
-        expect(updatedData[0]).toStrictEqual({
-            Int: null,
-            Bool: null,
-            Data: null,
-            Date: null,
-            RowId: '448',
-        });
-        expect(updatedData[1]).toStrictEqual({
-            Int: null,
-            Bool: null,
-            AndAgain: null,
-            RowId: '447',
-        });
-    });
-
-    test('edited row changed some values', () => {
-        const updatedData = getUpdatedDataFromGrid(
-            originalData,
-            [
-                Map<string, any>({
-                    RowId: '448',
-                    Value: null,
-                    Data: undefined,
-                    AndAgain: 'again',
-                    Name: 'S-20190516-9042',
-                    Other: 'other1',
-                    Bool: '',
-                    Int: '',
-                    Date: '2021-12-23 14:34',
-                }),
-                Map<string, any>({
-                    RowId: '447',
-                    Value: '447 Value',
-                    Data: 'data1',
-                    AndAgain: null,
-                    Name: 'S-20190516-4622',
-                    Other: 'other2',
-                    Bool: '',
-                    Int: '0',
-                    Date: null,
-                }),
-                Map<string, any>({
-                    RowId: '446',
-                    Value: 'new val',
-                    Data: 'data1',
-                    AndAgain: 'change me',
-                    Name: 'S-20190516-2368',
-                    Other: 'other3',
-                    Bool: false,
-                    Int: 66,
-                    Date: '1922-08-21 00:00',
-                }),
-                Map<string, any>({
-                    RowId: '445',
-                    Value: 'val',
-                    Data: 'other data',
-                    AndAgain: 'again',
-                    Name: 'S-20190516-9512',
-                    Other: null,
-                    Bool: true,
-                    Int: 5,
-                    Date: null,
-                }),
-            ],
-            'RowId',
-            queryInfo
-        );
-        expect(updatedData).toHaveLength(4);
-        expect(updatedData[0]).toStrictEqual({
-            Int: null,
-            Bool: null,
-            Data: null,
-            Date: '2021-12-23 14:34',
-            RowId: '448',
-        });
-        expect(updatedData[1]).toStrictEqual({
-            Int: 0,
-            Bool: null,
-            AndAgain: null,
-            Value: '447 Value',
-            RowId: '447',
-        });
-        expect(updatedData[2]).toStrictEqual({
-            Int: 66,
-            Bool: false,
-            Value: 'new val',
-            AndAgain: 'change me',
-            RowId: '446',
-        });
-        expect(updatedData[3]).toStrictEqual({
-            Bool: true,
-            Data: 'other data',
-            RowId: '445',
-        });
-    });
-
-    test('edited row added field', () => {
-        const updatedData = getUpdatedDataFromGrid(
-            originalData,
-            [
-                Map<string, any>({
-                    RowId: '448',
-                    'New Field': 'new value',
-                    Bool2: false,
-                    Int2: 22,
-                }),
-            ],
-            'RowId',
-            queryInfo
-        );
-        expect(updatedData).toHaveLength(1);
-        expect(updatedData[0]).toStrictEqual({
-            'New Field': 'new value',
-            Bool2: false,
-            Int2: 22,
-            RowId: '448',
-        });
-    });
-
-    test('with altUpdateKeys', () => {
-        const updatedData = getUpdatedDataFromGrid(
-            originalData,
-            [
-                Map<string, any>({
-                    RowId: '448',
-                    Data: 'data1',
-                    'New Field': 'new value',
-                    Bool2: false,
-                    Int2: 22,
-                }),
-            ],
-            'RowId',
-            queryInfoWithAltPK
-        );
-        expect(updatedData).toHaveLength(1);
-        expect(updatedData[0]).toStrictEqual({
-            'New Field': 'new value',
-            Data: 'data1',
-            Bool2: false,
-            Int2: 22,
-            RowId: '448',
-        });
-    });
-
-    test('row added field but no value', () => {
-        const updatedData = getUpdatedDataFromGrid(
-            originalData,
-            [
-                Map<string, any>({
-                    RowId: '448',
-                    'New Field': '',
-                    Bool: true,
-                    Int: 0,
-                }),
-            ],
-            'RowId',
-            queryInfo
-        );
-        expect(updatedData).toHaveLength(0);
     });
 });
